requirements.txt
google-adk==1.3.0
fastapi
uvicorn
pg8000
python-dotenv
opentelemetry-sdk
google-cloud-logging
pytest
pytest-asyncio          # for async FastAPI endpoints
httpx                   # lightweight HTTP client used by FastAPI TestClient
pandas 
numpy